pyproject.toml
[tool.poetry]
name = "five-year-journal"
version = "0.1.0"
description = ""
authors = ["Patrick Rodrigues <patrick.pwall@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.2"
uvicorn = "^0.17.6"
sqlmodel = "^0.0.6"
asyncpg = "^0.25.0"
PyJWT = "^2.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = "1.8.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = {extras = ["d"], version = "^22.3.0"}
python-dotenv = "^0.20.0"
aiosqlite = "^0.17.0"
httpx = "^0.22.0"
pytest-asyncio = "^0.18.3"
isort = "^5.10.1"
mypy = "^0.950"
flake8 = "^4.0.1"
pytest-cov = "^3.0.0"
pre-commit = "^2.18.1"
Faker = "^13.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = [
    "tests",
]
addopts = "-vv --cov-report=term-missing --cov=five_year_journal --cov-branch --cov-report=html"

[tool.coverage.report]
omit = [
    "five_year_journal/__main__.py",
    "five_year_journal/api/__init__.py",
]

[tool.black]
line-length = 79
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_local_folder = ["tests", "five_year_journal"]

[tool.mypy]
files = "five_year_journal"
ignore_missing_imports = true
plugins = "pydantic.mypy"
